+++ src/pages/Chatting/index.tsx
@@ -1,10 +1,10 @@
 import { useParams } from "react-router-dom";
 
-import { FullChat } from "components/Chatting";
+import Chat from "components/Chat";
 
 const Chatting = () => {
   const { roomId } = useParams() as { roomId: string };
-  return <FullChat roomId={roomId} />;
+  return <Chat roomId={roomId} layoutType="fullchat" />;
 };
 
 export default Chatting;
